requirements.txt
elasticsearch==6.3.1
elasticsearch-dsl==6.2.1
Pillow==5.3.0
python-geohash==0.8.5
flask==1.0.2
uwsgi==2.0.17.1
uwsgidecorators==1.1.0
colour==0.1.5
